refactor: move agri models to aides...
... since they are actually linked to Aides
