+++ apps/aides/admin.py
@@ -9,6 +9,10 @@ from .models import (
     Programme,
     Organisme,
     ZoneGeographique,
+    Filiere,
+    SousFiliere,
+    Production,
+    GroupementProducteurs,
     Aide,
 )
 
@@ -77,6 +81,51 @@ class ZoneGeographiqueAdmin(AbstractGristModelAdmin):
     list_filter = ("type",)
     list_select_related = ("parent", "epci")
     fields = ("parent", "type", "nom", "epci")
+
+
+@admin.register(GroupementProducteurs)
+class GroupementProducteursAdmin(AbstractGristModelAdmin):
+    list_display = AbstractGristModelAdmin.list_display + (
+        "nom",
+        "libelle",
+        "is_real",
+    )
+    list_display_links = AbstractGristModelAdmin.list_display_links + ("nom",)
+    fields = ("nom", "libelle")
+
+
+@admin.register(Filiere)
+class FiliereAdmin(AbstractGristModelAdmin):
+    list_display = AbstractGristModelAdmin.list_display + (
+        "nom",
+        "position",
+    )
+    list_display_links = AbstractGristModelAdmin.list_display_links + ("nom",)
+    fields = ("nom", "position")
+
+
+@admin.register(SousFiliere)
+class SousFiliereAdmin(AbstractGristModelAdmin):
+    list_display = AbstractGristModelAdmin.list_display + (
+        "nom",
+        "filiere",
+    )
+    list_display_links = AbstractGristModelAdmin.list_display_links + ("nom",)
+    list_filter = ("filiere",)
+    list_select_related = ("filiere",)
+    fields = ("nom", "filiere")
+
+
+@admin.register(Production)
+class ProductionAdmin(AbstractGristModelAdmin):
+    list_display = AbstractGristModelAdmin.list_display + (
+        "nom",
+        "sous_filiere",
+    )
+    list_display_links = AbstractGristModelAdmin.list_display_links + ("nom",)
+    list_filter = ("sous_filiere",)
+    list_select_related = ("sous_filiere",)
+    fields = ("nom", "sous_filiere")
 
 
 @admin.register(Aide)
