+++ conf/settings/default.py
@@ -5,4 +5,4 @@ from .apps.matomo import *  # noqa
 from .apps.sentry import *  # noqa
 
 from .apps.agri import *  # noqa
-from .apps.grist import *  # noqa
+from .apps.aides import *  # noqa
